Refactoring Vote: service, controller
In the service, I removed the 'get' method and added the 'getTodayByUser' method.
In the 'update' method, within the 'hasVoteTodayByUser', I added a check for the existence of a vote for a restaurant.
In the controller, I removed the 'get' method and refactored 'update/create' to pass only the restaurant ID as a parameter.
In the controller's tests, I added the 'updateDuplicate' method and made some minor adjustments.
I removed unnecessary test data.