Refactoring restaurant: repos, service, contr
Added the getAllWithMenu method to the repository and service to retrieve all restaurants with their menus for a specific date.
In the user controller, modified the endpoints, so now the user receives restaurants with menus for today.
In the admin controller, updated the endpoints, added two methods to retrieve restaurants with menus for today and a specific date.
 Added tests for the new methods.